+++ examples/diff3.pl
