+++ examples/moin_differ.rb
