+++ examples/penn_differ.py
