+++ lib/dyph3/support/merger.rb
