+++ lib/dyph3/two_way_differs/resig_diff.rb
@@ -1,7 +1,9 @@
 module Dyph3
   module TwoWayDiffers
-    class ResigDiff
-      def self.diff(old_text_array, new_text_array)
+    module ResigDiff
+      extend self
+
+      def diff(old_text_array, new_text_array)
         raise ArgumentError, "Argument is not an array." unless  old_text_array.is_a?(Array) && new_text_array.is_a?(Array)
 
         old_text, new_text = [old_text_array, new_text_array].map(&:dup)
@@ -18,7 +20,7 @@ module Dyph3
 
       private
 
-        def self.create_diffing_hash(values)
+        def create_diffing_hash(values)
           hash = {}
           values.each_with_index do |value, i|
             hash[value] ||= { rows: [] }
@@ -27,7 +29,7 @@ module Dyph3
           hash
         end
 
-        def self.find_single_matches(new_hash, old_hash, old_text, new_text)
+        def find_single_matches(new_hash, old_hash, old_text, new_text)
           new_hash.keys.each do |i|
             if (!old_hash[i].nil? && new_hash[i][:rows].length == 1 && old_hash[i][:rows].length == 1)
               new_hash_row = new_hash[i][:rows][0]
@@ -38,7 +40,7 @@ module Dyph3
           end
         end
 
-        def self.find_multi_matches(new_text, old_text, caller:)
+        def find_multi_matches(new_text, old_text, caller:)
           offset = get_offset(caller: caller)
           set_range(new_text, old_text, caller: caller).each do |i|
             if is_unchanged?(new_text, old_text, i, offset: offset, caller: caller)
@@ -49,7 +51,7 @@ module Dyph3
           end
         end
 
-        def self.get_offset(caller:)
+        def get_offset(caller:)
           if caller == :ascending
             1
           elsif caller == :descending
@@ -59,7 +61,7 @@ module Dyph3
           end
         end
 
-        def self.set_range(new_text, old_text, caller:)
+        def set_range(new_text, old_text, caller:)
           if caller == :ascending
             (0 ... new_text.length)
           elsif caller == :descending
@@ -70,7 +72,7 @@ module Dyph3
         end
 
 
-       def self.is_unchanged?(new_text, old_text, i, offset:, caller:)
+       def is_unchanged?(new_text, old_text, i, offset:, caller:)
           optional(new_text[i]).text.value &&                           # current value is marked as shared
           !(optional(new_text[i + offset]).text.value) &&               # value + offset is not marked as shared
           boundry_check(new_text, old_text, i, caller: caller) &&       # not off the end of the array
@@ -78,7 +80,7 @@ module Dyph3
           new_text[i + offset] == old_text[new_text[i].row + offset ]   # and the value in question matches
         end
 
-        def self.boundry_check(new_text, old_text, i, caller:)
+        def boundry_check(new_text, old_text, i, caller:)
           if caller == :ascending
             new_text[i].row + 1 < old_text.length
           elsif caller == :descending
@@ -88,7 +90,7 @@ module Dyph3
           end
         end
 
-        def self.optional(value)
+        def optional(value)
           Monads::Optional.new(value)
         end
       end
